Refactor HdfStorage out of tables.py
In order to support old and new PyTables versions, this
allows for choosing during service creation the storage
provider. This is the first step in to allowing non-HDF
backends.
